requirements.txt
Python 3.8.3
python_telegram_bot 12.8
